[feat]:#142 오피셜댓글-> 카테고리 상태변경

+++ src/main/java/ussum/homepage/infra/jpa/post/entity/BoardCode.java
@@ -30,7 +30,7 @@ public enum BoardCode {
                 .orElseThrow(() -> new InvalidValueException(INVALID_BOARDCODE));
     }
 
-    public static BoardCode getEnumBoardCodeFromBoardId(int boardId) {
+    public static BoardCode getEnumBoardCodeFromBoardId(Long boardId) {
         return Arrays.stream(values())
                 .filter(boardCode -> boardCode.boardId == boardId)
                 .findFirst()
